loader: getRawSegmentData() return size in uint64_t.
This size is casted to std::size_t by loader's users so it can be used
in Capstone's cs_disasm_iter().

+++ include/retdec/bin2llvmir/optimizations/decoder/decoder.h
@@ -53,7 +53,7 @@ class Decoder : public llvm::ModulePass
 				Abi* a);
 
 	private:
-		using ByteData = typename std::pair<const std::uint8_t*, std::uint64_t>;
+		using ByteData = typename std::pair<const std::uint8_t*, std::size_t>;
 
 	private:
 		bool runCatcher();
