+++ include/retdec/loader/loader/image.h
@@ -71,7 +71,7 @@ public:
 	const Segment* getSegmentFromAddress(std::uint64_t address) const;
 	const Segment* getEpSegment();
 
-	std::pair<const std::uint8_t*, std::size_t> getRawSegmentData(std::uint64_t address) const;
+	std::pair<const std::uint8_t*, std::uint64_t> getRawSegmentData(std::uint64_t address) const;
 
 	const std::string& getStatusMessage() const;
 	const retdec::fileformat::LoaderErrorInfo & getLoaderErrorInfo() const;
